Moved auto-submision to the search folder.
